Server ignores old requestvote response messages
When the network is slow and lossy, then there's a risk that we might
receive the requestvote response much later. So late that the term of
the server who sent the requestvote message doesn't match the
addressee's term. We now ignore these old responses to ensure the
election safety invariant.