Ignore RequestVote when leader is established. (#58)
This behavior is described in the original Raft paper and is required to
prevent removed nodes from disrupting the cluster.

This causes other problems though, as a split node may trigger election
which is rejected but leaves it with a future term.  The generally
accepted solution for this appears to be the use of PreVote RPC to avoid
election that can never succeed.

Signed-off-by: Li Wei <wei.g.li@intel.com>