DAOS-1876 build: Regenerate object files if headers change
Currently the Makefile recompiles object files only if the
corresponding source file has changed. We have run into a scenario
where a type definition was updated in a header causing different
object files to end up with multiple definitions of that same type
and resulting in memory corruption / segfaults.

Ideally, recompilation should occur only when an included header file
changes. But the simpler approach is to force recompilation when any
of the header files have changed; this works well for small projects.

Signed-off-by: Omkar Kulkarni <omkar.kulkarni@intel.com>