DAOS-1627 raft: Membership Changes
Instead of using flags for tracking the status of a node, it can be
interpreted from the most recent log entries affecting a node.  Making
repeated changes to the flags and keeping track of them proves to be
error prone when there are a series of membership changes. This
solution also renders a lot of code redundant, e.g. in raft_offer_log
and raft_pop_log and allows for a more elegant implementation.

Signed-off-by: Omkar Kulkarni <omkar.kulkarni@intel.com>